Added Local Notifications to example_projects
Added a small display of local notifications to our example project.

Note: current location where the local notification is being called will be changed as soon as we figure out the viewController. However, it is an example of how to request user authorization for notifications as well as display a local notification.